+++ Assets/Plugins/StreamChat/Core/DTO/Models/PolicyDTO.cs
@@ -17,7 +17,7 @@ namespace StreamChat.Core.DTO.Models
     internal partial class PolicyDTO
     {
         [Newtonsoft.Json.JsonProperty("action", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Action { get; set; }
+        public int? Action { get; set; }
 
         [Newtonsoft.Json.JsonProperty("created_at", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public System.DateTimeOffset? CreatedAt { get; set; }
@@ -29,7 +29,7 @@ namespace StreamChat.Core.DTO.Models
         public bool? Owner { get; set; }
 
         [Newtonsoft.Json.JsonProperty("priority", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Priority { get; set; }
+        public int? Priority { get; set; }
 
         [Newtonsoft.Json.JsonProperty("resources", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public System.Collections.Generic.List<string> Resources { get; set; }
